Change html and css to solve task
The html and css have been built in a way that they make use of the
tools available. Flex has been used mainly to build the layout.
Flex alignmenmt and CSS styling have been done in order to resemble
the example layout from the task as closely as possible. The image
shown is of own personal origin so no attribution is necessary.